src/requirements.txt
asgiref==3.2.3
blessings==1.7
bpython==0.18
certifi==2019.11.28
chardet==3.0.4
curtsies==0.3.1
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-countries==5.5
django-crispy-forms==1.8.1
django-extensions==2.2.8
greenlet==0.4.15
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
pkg-resources==0.0.0
Pygments==2.5.2
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
wcwidth==0.1.8
